// partner-search/frontend/src/app/api/model/group-dto.ts
/**
 * traefik-microstack API Gateway
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Partner group/association (Verbund) information
 */
export interface GroupDto { 
    /**
     * Unique group identifier
     */
    groupId?: number;
    /**
     * Group type/category
     */
    groupType?: number;
    /**
     * Advisor number
     */
    advisorNumber?: string;
    /**
     * Alpha code
     */
    alphaCode?: string;
    /**
     * Group name (part 1)
     */
    name1?: string;
    /**
     * Group name (part 2)
     */
    name2?: string;
    /**
     * Group name (part 3)
     */
    name3?: string;
    /**
     * Postal code
     */
    postalCode?: string;
    /**
     * City
     */
    city?: string;
    /**
     * Legal form
     */
    legalForm?: number;
    /**
     * Main industry
     */
    mainIndustry?: number;
    /**
     * Last modification date
     */
    modificationDate?: string;
    /**
     * Modifier ID
     */
    modifierId?: string;
}