// partner-search/frontend/src/app/api/model/group-list-response.ts
/**
 * traefik-microstack API Gateway
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupDto } from './group-dto';


/**
 * Group list response containing all partner groups/associations
 */
export interface GroupListResponse { 
    /**
     * List of all partner groups/associations
     */
    groups: Array<GroupDto>;
    /**
     * Total number of groups
     */
    totalCount?: number;
    /**
     * Number of groups returned in this response
     */
    returnedCount?: number;
}